+++ autom4te.cache/traces.1
@@ -20,7 +20,7 @@ m4trace:aclocal.m4:779: -1- m4_include([m4/pkg.m4])
 m4trace:aclocal.m4:780: -1- m4_include([m4/shapelib.m4])
 m4trace:aclocal.m4:781: -1- m4_include([m4/udunits.m4])
 m4trace:aclocal.m4:782: -1- m4_include([acinclude.m4])
-m4trace:configure.ac:12: -1- AC_INIT([GrADS], [2.2.dev], [jadams21@gmu.edu])
+m4trace:configure.ac:12: -1- AC_INIT([GrADS], [2.2.2], [jadams21@gmu.edu])
 m4trace:configure.ac:12: -1- m4_pattern_forbid([^_?A[CHUM]_])
 m4trace:configure.ac:12: -1- m4_pattern_forbid([_AC_])
 m4trace:configure.ac:12: -1- m4_pattern_forbid([^LIBOBJS$], [do not use LIBOBJS directly, use AC_LIBOBJ (see section `AC_LIBOBJ vs LIBOBJS'])
