+++ configure
@@ -1,6 +1,6 @@
 #! /bin/sh
 # Guess values for system-dependent variables and create Makefiles.
-# Generated by GNU Autoconf 2.69 for GrADS 2.2.dev.
+# Generated by GNU Autoconf 2.69 for GrADS 2.2.2.
 #
 # Report bugs to <jadams21@gmu.edu>.
 #
@@ -590,8 +590,8 @@ MAKEFLAGS=
 # Identity of this package.
 PACKAGE_NAME='GrADS'
 PACKAGE_TARNAME='grads'
-PACKAGE_VERSION='2.2.dev'
-PACKAGE_STRING='GrADS 2.2.dev'
+PACKAGE_VERSION='2.2.2'
+PACKAGE_STRING='GrADS 2.2.2'
 PACKAGE_BUGREPORT='jadams21@gmu.edu'
 PACKAGE_URL=''
 
@@ -1425,7 +1425,7 @@ if test "$ac_init_help" = "long"; then
   # Omit some internal or obsolete options to make the list less imposing.
   # This message is too long to be a string in the A/UX 3.1 sh.
   cat <<_ACEOF
-\`configure' configures GrADS 2.2.dev to adapt to many kinds of systems.
+\`configure' configures GrADS 2.2.2 to adapt to many kinds of systems.
 
 Usage: $0 [OPTION]... [VAR=VALUE]...
 
@@ -1499,7 +1499,7 @@ fi
 
 if test -n "$ac_init_help"; then
   case $ac_init_help in
-     short | recursive ) echo "Configuration of GrADS 2.2.dev:";;
+     short | recursive ) echo "Configuration of GrADS 2.2.2:";;
    esac
   cat <<\_ACEOF
 
@@ -1650,7 +1650,7 @@ fi
 test -n "$ac_init_help" && exit $ac_status
 if $ac_init_version; then
   cat <<\_ACEOF
-GrADS configure 2.2.dev
+GrADS configure 2.2.2
 generated by GNU Autoconf 2.69
 
 Copyright (C) 2012 Free Software Foundation, Inc.
@@ -2194,7 +2194,7 @@ cat >config.log <<_ACEOF
 This file contains any messages produced by compilers while
 running configure, to aid debugging if configure makes a mistake.
 
-It was created by GrADS $as_me 2.2.dev, which was
+It was created by GrADS $as_me 2.2.2, which was
 generated by GNU Autoconf 2.69.  Invocation command line was
 
   $ $0 $@
@@ -3064,7 +3064,7 @@ fi
 
 # Define the identity of the package.
  PACKAGE='grads'
- VERSION='2.2.dev'
+ VERSION='2.2.2'
 
 
 # Some tools Automake needs.
@@ -24300,7 +24300,7 @@ fi
 
 # Some non-configuration-dependent GrADS macros
 
-$as_echo "#define GRADS_VERSION \"2.2.dev\"" >>confdefs.h
+$as_echo "#define GRADS_VERSION \"2.2.2\"" >>confdefs.h
 
 
 # These are used to add the necessary source files to Makefile targets
@@ -24385,7 +24385,7 @@ fi
 # Write output files
 # ----------------------------------------------------
 echo "configure: creating src/VERSION"
-echo 2.2.dev > src/VERSION
+echo 2.2.2 > src/VERSION
 
 echo "configure: creating src/buildinfo.h"
 build_info="Configured on `date +%D` for ${host}"
@@ -24957,7 +24957,7 @@ cat >>$CONFIG_STATUS <<\_ACEOF || ac_write_fail=1
 # report actual input values of CONFIG_FILES etc. instead of their
 # values after options handling.
 ac_log="
-This file was extended by GrADS $as_me 2.2.dev, which was
+This file was extended by GrADS $as_me 2.2.2, which was
 generated by GNU Autoconf 2.69.  Invocation command line was
 
   CONFIG_FILES    = $CONFIG_FILES
@@ -25023,7 +25023,7 @@ _ACEOF
 cat >>$CONFIG_STATUS <<_ACEOF || ac_write_fail=1
 ac_cs_config="`$as_echo "$ac_configure_args" | sed 's/^ //; s/[\\""\`\$]/\\\\&/g'`"
 ac_cs_version="\\
-GrADS config.status 2.2.dev
+GrADS config.status 2.2.2
 configured by $0, generated by GNU Autoconf 2.69,
   with options \\"\$ac_cs_config\\"
 
@@ -26887,7 +26887,7 @@ fi
    echo
    echo \
 "  +=========================================================================+"
-   echo "   GrADS "2.2.dev ":" $build_info
+   echo "   GrADS "2.2.2 ":" $build_info
    echo \
 "  +=========================================================================+"
    echo
