+++ src/VERSION
@@ -1,1 +1,1 @@
-2.2.dev
+2.2.2
