+++ src/setup.py
@@ -1,3 +1,3 @@
 from distutils.core import setup, Extension
-setup(name='gradspy', version='1.0',  \
+setup(name='gradspy', version='1.1',  \
       ext_modules=[Extension('gradspy', ['gradspy.c'])])
